requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pyserial>=3.5
boto3>=1.28.0
tweepy>=4.14.0
google-generativeai>=0.4.0
requests>=2.31.0
python-dotenv>=1.0.0
argopy>=0.1.14
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
marimo>=0.3.0
altair>=5.0.0
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
